비회원신고기능감추기, 게시글 삽입 수정오류정정 등

+++ Devlog/src/main/java/com/devlog/project/board/freeboard/controller/FreeboardController.java
@@ -65,7 +65,8 @@ public class FreeboardController {
 	// 게시글 상세조회
 	@GetMapping("/freeboard/{boardNo}")
 	public String selectFreeboardDetail( 
-			@PathVariable("boardNo") int boardNo
+			//@PathVariable("boardNo") int boardNo
+			@PathVariable("boardNo") Long boardNo
 			, Model model 
 			, RedirectAttributes ra 
 			, @SessionAttribute(value = "loginMember", required=false) MemberLoginResponseDTO loginMember
